+++ bricks/flavors/__brick__/main_{{#flavors}}{{{.}}}{{/flavors}}.dart
@@ -1,0 +1,3 @@
+void main() {
+  {{#flavors}}print('Running in {{.}} mode...');{{/flavors}}
+}