+++ packages/mason_logger/test/src/mason_logger_test.dart
@@ -11,10 +11,12 @@ void main() {
   group('Logger', () {
     late Stdout stdout;
     late Stdin stdin;
+    late Stdout stderr;
 
     setUp(() {
       stdout = MockStdout();
       stdin = MockStdin();
+      stderr = MockStdout();
     });
 
     group('.write', () {
@@ -79,14 +81,14 @@ void main() {
     });
 
     group('.err', () {
-      test('writes line to stdout', () {
+      test('writes line to stderr', () {
         StdioOverrides.runZoned(
           () {
             const message = 'test message';
             Logger().err(message);
-            verify(() => stdout.writeln(lightRed.wrap(message))).called(1);
+            verify(() => stderr.writeln(lightRed.wrap(message))).called(1);
           },
-          stdout: () => stdout,
+          stderr: () => stderr,
         );
       });
     });
@@ -120,33 +122,33 @@ void main() {
     });
 
     group('.warn', () {
-      test('writes line to stdout', () {
+      test('writes line to stderr', () {
         StdioOverrides.runZoned(
           () {
             const message = 'test message';
             Logger().warn(message);
             verify(
               () {
-                stdout.writeln(yellow.wrap(styleBold.wrap('[WARN] $message')));
+                stderr.writeln(yellow.wrap(styleBold.wrap('[WARN] $message')));
               },
             ).called(1);
           },
-          stdout: () => stdout,
+          stderr: () => stderr,
         );
       });
 
-      test('writes line to stdout with custom tag', () {
+      test('writes line to stderr with custom tag', () {
         StdioOverrides.runZoned(
           () {
             const message = 'test message';
             Logger().warn(message, tag: '🚨');
             verify(
               () {
-                stdout.writeln(yellow.wrap(styleBold.wrap('[🚨] $message')));
+                stderr.writeln(yellow.wrap(styleBold.wrap('[🚨] $message')));
               },
             ).called(1);
           },
-          stdout: () => stdout,
+          stderr: () => stderr,
         );
       });
     });
@@ -416,6 +418,378 @@ void main() {
       });
     });
 
+    group('.chooseAny', () {
+      test(
+          'enter selects the nothing '
+          'when defaultValues is not specified.', () {
+        StdioOverrides.runZoned(
+          () {
+            const message = 'test message';
+            when(() => stdin.readByteSync()).thenReturn(10);
+            final actual = Logger().chooseAny(
+              message,
+              choices: ['a', 'b', 'c'],
+            );
+            expect(actual, isEmpty);
+            verifyInOrder([
+              () => stdout.write('\x1b7'),
+              () => stdout.write('\x1b[?25l'),
+              () => stdout.writeln(message),
+              () => stdout.write(green.wrap('❯')),
+              () => stdout.write(' ◯  ${lightCyan.wrap('a')}'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  b'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  c'),
+            ]);
+          },
+          stdout: () => stdout,
+          stdin: () => stdin,
+        );
+      });
+
+      test('enter selects the default values when specified.', () {
+        StdioOverrides.runZoned(
+          () {
+            const message = 'test message';
+            const expected = ['b', 'c'];
+            when(() => stdin.readByteSync()).thenReturn(10);
+            final actual = Logger().chooseAny(
+              message,
+              choices: ['a', 'b', 'c'],
+              defaultValues: ['b', 'c'],
+            );
+            expect(actual, equals(expected));
+            verifyInOrder([
+              () => stdout.write('\x1b7'),
+              () => stdout.write('\x1b[?25l'),
+              () => stdout.writeln(message),
+              () => stdout.write(green.wrap('❯')),
+              () => stdout.write(' ◯  a'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout
+                  .write(' ${lightCyan.wrap('◉')}  ${lightCyan.wrap('b')}'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout
+                  .write(' ${lightCyan.wrap('◉')}  ${lightCyan.wrap('c')}'),
+            ]);
+          },
+          stdout: () => stdout,
+          stdin: () => stdin,
+        );
+      });
+
+      test('space selected/deselects the values.', () {
+        StdioOverrides.runZoned(
+          () {
+            const message = 'test message';
+            const expected = ['b', 'c'];
+            final bytes = [32, 32, 27, 91, 66, 32, 27, 91, 66, 32, 10];
+            when(() => stdin.readByteSync()).thenAnswer((_) {
+              return bytes.removeAt(0);
+            });
+            final actual = Logger().chooseAny(
+              message,
+              choices: ['a', 'b', 'c'],
+            );
+            expect(actual, equals(expected));
+            verifyInOrder([
+              () => stdout.write('\x1b7'),
+              () => stdout.write('\x1b[?25l'),
+              () => stdout.writeln(message),
+              () => stdout.write(green.wrap('❯')),
+              () => stdout
+                  .write(' ${lightCyan.wrap('◯')}  ${lightCyan.wrap('a')}'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  b'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  c'),
+              () => stdout.write('\x1b7'),
+              () => stdout.write('\x1b[?25l'),
+              () => stdout.writeln(message),
+              () => stdout.write(green.wrap('❯')),
+              () => stdout
+                  .write(' ${lightCyan.wrap('◉')}  ${lightCyan.wrap('a')}'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  b'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  c'),
+              () => stdout.write('\x1b7'),
+              () => stdout.write('\x1b[?25l'),
+              () => stdout.writeln(message),
+              () => stdout.write(green.wrap('❯')),
+              () => stdout
+                  .write(' ${lightCyan.wrap('◯')}  ${lightCyan.wrap('a')}'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  b'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  c'),
+              () => stdout.write('\x1b7'),
+              () => stdout.write('\x1b[?25l'),
+              () => stdout.writeln(message),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  a'),
+              () => stdout.write(green.wrap('❯')),
+              () => stdout
+                  .write(' ${lightCyan.wrap('◯')}  ${lightCyan.wrap('b')}'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  c'),
+              () => stdout.write('\x1b7'),
+              () => stdout.write('\x1b[?25l'),
+              () => stdout.writeln(message),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  a'),
+              () => stdout.write(green.wrap('❯')),
+              () => stdout
+                  .write(' ${lightCyan.wrap('◉')}  ${lightCyan.wrap('b')}'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  c'),
+              () => stdout.write('\x1b7'),
+              () => stdout.write('\x1b[?25l'),
+              () => stdout.writeln(message),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  a'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout
+                  .write(' ${lightCyan.wrap('◉')}  ${lightCyan.wrap('b')}'),
+              () => stdout.write(green.wrap('❯')),
+              () => stdout.write(' ◯  ${lightCyan.wrap('c')}'),
+              () => stdout.write('\x1b7'),
+              () => stdout.write('\x1b[?25l'),
+              () => stdout.writeln(message),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  a'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout
+                  .write(' ${lightCyan.wrap('◉')}  ${lightCyan.wrap('b')}'),
+              () => stdout.write(green.wrap('❯')),
+              () => stdout
+                  .write(' ${lightCyan.wrap('◉')}  ${lightCyan.wrap('c')}'),
+            ]);
+          },
+          stdout: () => stdout,
+          stdin: () => stdin,
+        );
+      });
+
+      test('down arrow selects next index', () {
+        StdioOverrides.runZoned(
+          () {
+            const message = 'test message';
+            final bytes = [27, 91, 66, 10];
+            when(() => stdin.readByteSync()).thenAnswer((_) {
+              return bytes.removeAt(0);
+            });
+            final actual = Logger().chooseAny(
+              message,
+              choices: ['a', 'b', 'c'],
+            );
+            expect(actual, equals(isEmpty));
+            verifyInOrder([
+              () => stdout.write('\x1b7'),
+              () => stdout.write('\x1b[?25l'),
+              () => stdout.writeln(message),
+              () => stdout.write(green.wrap('❯')),
+              () => stdout
+                  .write(' ${lightCyan.wrap('◯')}  ${lightCyan.wrap('a')}'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  b'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  c'),
+              () => stdout.write('\x1b7'),
+              () => stdout.write('\x1b[?25l'),
+              () => stdout.writeln(message),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  a'),
+              () => stdout.write(green.wrap('❯')),
+              () => stdout
+                  .write(' ${lightCyan.wrap('◯')}  ${lightCyan.wrap('b')}'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  c'),
+            ]);
+          },
+          stdout: () => stdout,
+          stdin: () => stdin,
+        );
+      });
+
+      test('k selects next index', () {
+        StdioOverrides.runZoned(
+          () {
+            const message = 'test message';
+            final bytes = [107, 10];
+            when(() => stdin.readByteSync()).thenAnswer((_) {
+              return bytes.removeAt(0);
+            });
+            final actual = Logger().chooseAny(
+              message,
+              choices: ['a', 'b', 'c'],
+            );
+            expect(actual, equals(isEmpty));
+            verifyInOrder([
+              () => stdout.write('\x1b7'),
+              () => stdout.write('\x1b[?25l'),
+              () => stdout.writeln(message),
+              () => stdout.write(green.wrap('❯')),
+              () => stdout
+                  .write(' ${lightCyan.wrap('◯')}  ${lightCyan.wrap('a')}'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  b'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  c'),
+              () => stdout.write('\x1b7'),
+              () => stdout.write('\x1b[?25l'),
+              () => stdout.writeln(message),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  a'),
+              () => stdout.write(green.wrap('❯')),
+              () => stdout
+                  .write(' ${lightCyan.wrap('◯')}  ${lightCyan.wrap('b')}'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  c'),
+            ]);
+          },
+          stdout: () => stdout,
+          stdin: () => stdin,
+        );
+      });
+
+      test('up arrow wraps to end', () {
+        StdioOverrides.runZoned(
+          () {
+            const message = 'test message';
+            final bytes = [27, 91, 65, 10];
+            when(() => stdin.readByteSync()).thenAnswer((_) {
+              return bytes.removeAt(0);
+            });
+            final actual = Logger().chooseAny(
+              message,
+              choices: ['a', 'b', 'c'],
+            );
+            expect(actual, isEmpty);
+            verifyInOrder([
+              () => stdout.write('\x1b7'),
+              () => stdout.write('\x1b[?25l'),
+              () => stdout.writeln(message),
+              () => stdout.write(green.wrap('❯')),
+              () => stdout.write(' ◯  ${lightCyan.wrap('a')}'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  b'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  c'),
+              () => stdout.write('\x1b7'),
+              () => stdout.write('\x1b[?25l'),
+              () => stdout.writeln(message),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  a'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  b'),
+              () => stdout.write(green.wrap('❯')),
+              () => stdout.write(' ◯  ${lightCyan.wrap('c')}'),
+            ]);
+          },
+          stdout: () => stdout,
+          stdin: () => stdin,
+        );
+      });
+
+      test('k wraps to end', () {
+        StdioOverrides.runZoned(
+          () {
+            const message = 'test message';
+            final bytes = [106, 10];
+            when(() => stdin.readByteSync()).thenAnswer((_) {
+              return bytes.removeAt(0);
+            });
+            final actual = Logger().chooseAny(
+              message,
+              choices: ['a', 'b', 'c'],
+            );
+            expect(actual, isEmpty);
+            verifyInOrder([
+              () => stdout.write('\x1b7'),
+              () => stdout.write('\x1b[?25l'),
+              () => stdout.writeln(message),
+              () => stdout.write(green.wrap('❯')),
+              () => stdout.write(' ◯  ${lightCyan.wrap('a')}'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  b'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  c'),
+              () => stdout.write('\x1b7'),
+              () => stdout.write('\x1b[?25l'),
+              () => stdout.writeln(message),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  a'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  b'),
+              () => stdout.write(green.wrap('❯')),
+              () => stdout.write(' ◯  ${lightCyan.wrap('c')}'),
+            ]);
+          },
+          stdout: () => stdout,
+          stdin: () => stdin,
+        );
+      });
+
+      test('down arrow wraps to beginning', () {
+        StdioOverrides.runZoned(
+          () {
+            const message = 'test message';
+            final bytes = [27, 91, 66, 27, 91, 66, 27, 91, 66, 10];
+            when(() => stdin.readByteSync()).thenAnswer((_) {
+              return bytes.removeAt(0);
+            });
+            final actual = Logger().chooseAny(
+              message,
+              choices: ['a', 'b', 'c'],
+            );
+            expect(actual, isEmpty);
+            verifyInOrder([
+              () => stdout.write('\x1b7'),
+              () => stdout.write('\x1b[?25l'),
+              () => stdout.writeln(message),
+              () => stdout.write(green.wrap('❯')),
+              () => stdout.write(' ◯  ${lightCyan.wrap('a')}'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  b'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  c'),
+              () => stdout.write('\x1b7'),
+              () => stdout.write('\x1b[?25l'),
+              () => stdout.writeln(message),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  a'),
+              () => stdout.write(green.wrap('❯')),
+              () => stdout.write(' ◯  ${lightCyan.wrap('b')}'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  c'),
+              () => stdout.write('\x1b7'),
+              () => stdout.write('\x1b[?25l'),
+              () => stdout.writeln(message),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  a'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  b'),
+              () => stdout.write(green.wrap('❯')),
+              () => stdout.write(' ◯  ${lightCyan.wrap('c')}'),
+              () => stdout.write('\x1b7'),
+              () => stdout.write('\x1b[?25l'),
+              () => stdout.writeln(message),
+              () => stdout.write(green.wrap('❯')),
+              () => stdout.write(' ◯  ${lightCyan.wrap('a')}'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  b'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  c'),
+            ]);
+          },
+          stdout: () => stdout,
+          stdin: () => stdin,
+        );
+      });
+    });
+
     group('.chooseOne', () {
       test(
           'enter selects the initial value '
@@ -436,11 +810,11 @@ void main() {
               () => stdout.writeln(message),
               () => stdout.write(green.wrap('❯')),
               () => stdout
-                  .write(' ${lightCyan.wrap('◉')} ${lightCyan.wrap('a')}'),
+                  .write(' ${lightCyan.wrap('◉')}  ${lightCyan.wrap('a')}'),
               () => stdout.write(green.wrap(' ')),
-              () => stdout.write(' ◯ b'),
+              () => stdout.write(' ◯  b'),
               () => stdout.write(green.wrap(' ')),
-              () => stdout.write(' ◯ c'),
+              () => stdout.write(' ◯  c'),
             ]);
           },
           stdout: () => stdout,
@@ -465,12 +839,42 @@ void main() {
               () => stdout.write('\x1b[?25l'),
               () => stdout.writeln(message),
               () => stdout.write(green.wrap(' ')),
-              () => stdout.write(' ◯ a'),
+              () => stdout.write(' ◯  a'),
               () => stdout.write(green.wrap('❯')),
               () => stdout
-                  .write(' ${lightCyan.wrap('◉')} ${lightCyan.wrap('b')}'),
+                  .write(' ${lightCyan.wrap('◉')}  ${lightCyan.wrap('b')}'),
               () => stdout.write(green.wrap(' ')),
-              () => stdout.write(' ◯ c'),
+              () => stdout.write(' ◯  c'),
+            ]);
+          },
+          stdout: () => stdout,
+          stdin: () => stdin,
+        );
+      });
+
+      test('space selects the default value when specified.', () {
+        StdioOverrides.runZoned(
+          () {
+            const message = 'test message';
+            const expected = 'b';
+            when(() => stdin.readByteSync()).thenReturn(32);
+            final actual = Logger().chooseOne(
+              message,
+              choices: ['a', 'b', 'c'],
+              defaultValue: 'b',
+            );
+            expect(actual, equals(expected));
+            verifyInOrder([
+              () => stdout.write('\x1b7'),
+              () => stdout.write('\x1b[?25l'),
+              () => stdout.writeln(message),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  a'),
+              () => stdout.write(green.wrap('❯')),
+              () => stdout
+                  .write(' ${lightCyan.wrap('◉')}  ${lightCyan.wrap('b')}'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  c'),
             ]);
           },
           stdout: () => stdout,
@@ -498,21 +902,21 @@ void main() {
               () => stdout.writeln(message),
               () => stdout.write(green.wrap('❯')),
               () => stdout
-                  .write(' ${lightCyan.wrap('◉')} ${lightCyan.wrap('a')}'),
+                  .write(' ${lightCyan.wrap('◉')}  ${lightCyan.wrap('a')}'),
               () => stdout.write(green.wrap(' ')),
-              () => stdout.write(' ◯ b'),
+              () => stdout.write(' ◯  b'),
               () => stdout.write(green.wrap(' ')),
-              () => stdout.write(' ◯ c'),
+              () => stdout.write(' ◯  c'),
               () => stdout.write('\x1b7'),
               () => stdout.write('\x1b[?25l'),
               () => stdout.writeln(message),
               () => stdout.write(green.wrap(' ')),
-              () => stdout.write(' ◯ a'),
+              () => stdout.write(' ◯  a'),
               () => stdout.write(green.wrap('❯')),
               () => stdout
-                  .write(' ${lightCyan.wrap('◉')} ${lightCyan.wrap('b')}'),
+                  .write(' ${lightCyan.wrap('◉')}  ${lightCyan.wrap('b')}'),
               () => stdout.write(green.wrap(' ')),
-              () => stdout.write(' ◯ c'),
+              () => stdout.write(' ◯  c'),
             ]);
           },
           stdout: () => stdout,
@@ -540,22 +944,22 @@ void main() {
               () => stdout.write('\x1b[?25l'),
               () => stdout.writeln(message),
               () => stdout.write(green.wrap(' ')),
-              () => stdout.write(' ◯ a'),
+              () => stdout.write(' ◯  a'),
               () => stdout.write(green.wrap('❯')),
               () => stdout
-                  .write(' ${lightCyan.wrap('◉')} ${lightCyan.wrap('b')}'),
+                  .write(' ${lightCyan.wrap('◉')}  ${lightCyan.wrap('b')}'),
               () => stdout.write(green.wrap(' ')),
-              () => stdout.write(' ◯ c'),
+              () => stdout.write(' ◯  c'),
               () => stdout.write('\x1b7'),
               () => stdout.write('\x1b[?25l'),
               () => stdout.writeln(message),
               () => stdout.write(green.wrap('❯')),
               () => stdout
-                  .write(' ${lightCyan.wrap('◉')} ${lightCyan.wrap('a')}'),
+                  .write(' ${lightCyan.wrap('◉')}  ${lightCyan.wrap('a')}'),
               () => stdout.write(green.wrap(' ')),
-              () => stdout.write(' ◯ b'),
+              () => stdout.write(' ◯  b'),
               () => stdout.write(green.wrap(' ')),
-              () => stdout.write(' ◯ c'),
+              () => stdout.write(' ◯  c'),
             ]);
           },
           stdout: () => stdout,
@@ -583,21 +987,21 @@ void main() {
               () => stdout.writeln(message),
               () => stdout.write(green.wrap('❯')),
               () => stdout
-                  .write(' ${lightCyan.wrap('◉')} ${lightCyan.wrap('a')}'),
+                  .write(' ${lightCyan.wrap('◉')}  ${lightCyan.wrap('a')}'),
               () => stdout.write(green.wrap(' ')),
-              () => stdout.write(' ◯ b'),
+              () => stdout.write(' ◯  b'),
               () => stdout.write(green.wrap(' ')),
-              () => stdout.write(' ◯ c'),
+              () => stdout.write(' ◯  c'),
               () => stdout.write('\x1b7'),
               () => stdout.write('\x1b[?25l'),
               () => stdout.writeln(message),
               () => stdout.write(green.wrap(' ')),
-              () => stdout.write(' ◯ a'),
+              () => stdout.write(' ◯  a'),
               () => stdout.write(green.wrap(' ')),
-              () => stdout.write(' ◯ b'),
+              () => stdout.write(' ◯  b'),
               () => stdout.write(green.wrap('❯')),
               () => stdout
-                  .write(' ${lightCyan.wrap('◉')} ${lightCyan.wrap('c')}'),
+                  .write(' ${lightCyan.wrap('◉')}  ${lightCyan.wrap('c')}'),
             ]);
           },
           stdout: () => stdout,
@@ -625,22 +1029,107 @@ void main() {
               () => stdout.write('\x1b[?25l'),
               () => stdout.writeln(message),
               () => stdout.write(green.wrap(' ')),
-              () => stdout.write(' ◯ a'),
+              () => stdout.write(' ◯  a'),
               () => stdout.write(green.wrap(' ')),
-              () => stdout.write(' ◯ b'),
+              () => stdout.write(' ◯  b'),
               () => stdout.write(green.wrap('❯')),
               () => stdout
-                  .write(' ${lightCyan.wrap('◉')} ${lightCyan.wrap('c')}'),
+                  .write(' ${lightCyan.wrap('◉')}  ${lightCyan.wrap('c')}'),
               () => stdout.write('\x1b7'),
               () => stdout.write('\x1b[?25l'),
               () => stdout.writeln(message),
               () => stdout.write(green.wrap('❯')),
               () => stdout
-                  .write(' ${lightCyan.wrap('◉')} ${lightCyan.wrap('a')}'),
+                  .write(' ${lightCyan.wrap('◉')}  ${lightCyan.wrap('a')}'),
               () => stdout.write(green.wrap(' ')),
-              () => stdout.write(' ◯ b'),
+              () => stdout.write(' ◯  b'),
               () => stdout.write(green.wrap(' ')),
-              () => stdout.write(' ◯ c'),
+              () => stdout.write(' ◯  c'),
+            ]);
+          },
+          stdout: () => stdout,
+          stdin: () => stdin,
+        );
+      });
+
+      test('j selects next index', () {
+        StdioOverrides.runZoned(
+          () {
+            const message = 'test message';
+            const expected = 'b';
+            final bytes = [107, 10];
+            when(() => stdin.readByteSync()).thenAnswer((_) {
+              return bytes.removeAt(0);
+            });
+            final actual = Logger().chooseOne(
+              message,
+              choices: ['a', 'b', 'c'],
+            );
+            expect(actual, equals(expected));
+            verifyInOrder([
+              () => stdout.write('\x1b7'),
+              () => stdout.write('\x1b[?25l'),
+              () => stdout.writeln(message),
+              () => stdout.write(green.wrap('❯')),
+              () => stdout
+                  .write(' ${lightCyan.wrap('◉')}  ${lightCyan.wrap('a')}'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  b'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  c'),
+              () => stdout.write('\x1b7'),
+              () => stdout.write('\x1b[?25l'),
+              () => stdout.writeln(message),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  a'),
+              () => stdout.write(green.wrap('❯')),
+              () => stdout
+                  .write(' ${lightCyan.wrap('◉')}  ${lightCyan.wrap('b')}'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  c'),
+            ]);
+          },
+          stdout: () => stdout,
+          stdin: () => stdin,
+        );
+      });
+
+      test('k selects previous index', () {
+        StdioOverrides.runZoned(
+          () {
+            const message = 'test message';
+            const expected = 'a';
+            final bytes = [106, 10];
+            when(() => stdin.readByteSync()).thenAnswer((_) {
+              return bytes.removeAt(0);
+            });
+            final actual = Logger().chooseOne(
+              message,
+              choices: ['a', 'b', 'c'],
+              defaultValue: 'b',
+            );
+            expect(actual, equals(expected));
+            verifyInOrder([
+              () => stdout.write('\x1b7'),
+              () => stdout.write('\x1b[?25l'),
+              () => stdout.writeln(message),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  a'),
+              () => stdout.write(green.wrap('❯')),
+              () => stdout
+                  .write(' ${lightCyan.wrap('◉')}  ${lightCyan.wrap('b')}'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  c'),
+              () => stdout.write('\x1b7'),
+              () => stdout.write('\x1b[?25l'),
+              () => stdout.writeln(message),
+              () => stdout.write(green.wrap('❯')),
+              () => stdout
+                  .write(' ${lightCyan.wrap('◉')}  ${lightCyan.wrap('a')}'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  b'),
+              () => stdout.write(green.wrap(' ')),
+              () => stdout.write(' ◯  c'),
             ]);
           },
           stdout: () => stdout,