+++ packages/mason_logger/CHANGELOG.md
@@ -1,3 +1,22 @@
+# 0.1.0-dev.12
+
+- fix: `chooseOne` API windows compatibility
+- feat: `chooseAny`
+  ```dart
+  /// Ask user to choose zero or more options.
+  final desserts = logger.chooseAny(
+    'Which desserts do you like?',
+    choices: ['🍦', '🍪', '🍩'],
+  );
+  ```
+
+# 0.1.0-dev.11
+
+- fix: write errors and warnings to `stderr`
+  - `Logger().err(...)`
+  - `Logger().warn(...)`
+  - `Logger().progress(...).fail(...)`
+
 # 0.1.0-dev.10
 
 - feat: `chooseOne` API